+++ udata_front/theme/gouvfr/assets/js/components/MarkdownEditor/Milkdown/LinkPreview/linkPreviewTooltip.ts
@@ -1,0 +1,9 @@
+import { TooltipProvider, tooltipFactory } from "@milkdown/plugin-tooltip";
+import { $ctx } from "@milkdown/utils";
+
+export const linkPreviewTooltipCtx = $ctx<TooltipProvider | null, "linkPreviewTooltip">(
+  null,
+  "linkPreviewTooltip"
+);
+
+export const linkPreviewTooltip = tooltipFactory("LINK_PREVIEW");