+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix release CI [#380](https://github.com/etalab/udata-front/pull/380)
 - Add stories to Resource component [#364](https://github.com/etalab/udata-front/pull/364)
 - Move Well to datagouv-components [#382](https://github.com/etalab/udata-front/pull/382)
+- Add markdown editor [#351](https://github.com/etalab/udata-front/pull/351)
 
 ## 3.5.4 (2024-03-20)